requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
psycopg[binary]
pydantic>=2.0
pydantic-settings
python-multipart
chromadb
httpx
pypdf
python-docx
openpyxl
python-pptx
